+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Release 5.5.1
+ * Fix DCQL Query extraction for SD-JWT
+ * Support more credential identifiers for presentation exchange
+ * Update to VC-K 5.5.1
+
 # Release 5.5.0
  * Android: Add NFC device engagement/retrieval and Bluetooth device retrieval mechanisms
  * Update to VC-K 5.5.0: Use OID4VCI Draft 15 for loading credentials